+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# dev
+- Add method to get a point cloud origin
 # 1.7.2
 - Add possibility to select extra dimensions to keep in standardization